// frontend/src/api/generated/trust-zones/trust-zones.ts
export type listTrustZonesApiV1OrganizationsMeZonesGetResponse200 = {
  data: TrustZoneRead[];
  status: 200;
};

export type listTrustZonesApiV1OrganizationsMeZonesGetResponse422 = {
  data: HTTPValidationError;
  status: 422;
};

export type listTrustZonesApiV1OrganizationsMeZonesGetResponseSuccess =
  listTrustZonesApiV1OrganizationsMeZonesGetResponse200 & {
    headers: Headers;
  };
export type listTrustZonesApiV1OrganizationsMeZonesGetResponseError =
  listTrustZonesApiV1OrganizationsMeZonesGetResponse422 & {
    headers: Headers;
  };

export type listTrustZonesApiV1OrganizationsMeZonesGetResponse =
  | listTrustZonesApiV1OrganizationsMeZonesGetResponseSuccess
  | listTrustZonesApiV1OrganizationsMeZonesGetResponseError;